requirements.txt
asgiref==3.5.0
astroid==2.11.3
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
colorama==0.4.4
dill==0.3.4
Django==4.0.4
django-appconf==1.0.5
django-ckeditor==6.4.0
django-imagekit==4.1.0
django-js-asset==2.0.0
gunicorn==20.1.0
isort==5.10.1
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
pilkit==2.0
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pylint==2.13.7
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
Werkzeug==2.1.1
whitenoise==6.0.0
wrapt==1.14.0
